Database Controller has been completed.
Database Controller has been completed... mostly. There are fields that
aren't currently handled in the gui part of the program so instead of
sending that data to the recipient controllers, they are printed to the
screen. This will be addressed in tomorrows meeting. I will have to
cancel my meeting with the client as I didn't realize how far we're
behind in the implementation of the product.