Added credential validation code
Added credential validation code in database integration module. Have
optimized the process to one method. Original intentions were to request
the username is valid, then request that the combination of the username
and password is valid. Since this code is entirely internal, it would be
quicker to request the username is valid and have it return the password
if it is, else invalid so that the handling module can streamline the
process.