Updated all icons to be 100x100 and <7kb
Updated all icons to be 100x100 and <7kb. By doing this I found out that
the lag has nothing to do with the file size, but the number of sql
statements to retrieve all the data. It's essentially innefficient.
Probably should change the statement to inner join on data so there's
just one pull at a request. This can be updated after the project
deadline.